Create Boruvka’s_algorithm.java
A spanning tree means all vertices must be connected. So the two disjoint subsets (discussed above) of vertices must be connected to make a Spanning Tree. And they must be connected with the minimum weight edge to make it a Minimum Spanning Tree.

Time Complexity of Boruvka’s algorithm is O(E log V) which is the same as Kruskal’s and Prim’s algorithms.
Boruvka’s algorithm is used as a step in a faster randomized algorithm that works in linear time O(E).
Boruvka’s algorithm is the oldest minimum spanning tree algorithm that was discovered by Boruvka in 1926, long before computers even existed. The algorithm was published as a method of constructing an efficient electricity network.
Space complexity: The space complexity of Boruvka’s algorithm is O(V).